Autogen A, MX, SRV (#3426)
