+++ integrationTest/helpers_integration_test.go
@@ -293,34 +293,6 @@ func withMeta(record *models.RecordConfig, metadata map[string]string) *models.R
 	return record
 }
 
-func a(name string, a string) *models.RecordConfig {
-	rdata, err := models.ParseA([]string{a}, "**current-domain**")
-	if err != nil {
-		panic(err)
-	}
-	return models.MustCreateRecord(name, rdata, nil, 300, "**current-domain**")
-}
-
-func mx(name string, preference uint16, mx string) *models.RecordConfig {
-	spreference := strconv.Itoa(int(preference))
-	rdata, err := models.ParseMX([]string{spreference, mx}, "**current-domain**")
-	if err != nil {
-		panic(err)
-	}
-	return models.MustCreateRecord(name, rdata, nil, 300, "**current-domain**")
-}
-
-func srv(name string, priority, weight, port uint16, target string) *models.RecordConfig {
-	spriority := strconv.Itoa(int(priority))
-	sweight := strconv.Itoa(int(weight))
-	sport := strconv.Itoa(int(port))
-	rdata, err := models.ParseSRV([]string{spriority, sweight, sport, target}, "**current-domain**")
-	if err != nil {
-		panic(err)
-	}
-	return models.MustCreateRecord(name, rdata, nil, 300, "**current-domain**")
-}
-
 func cfSingleRedirect(name string, code uint16, when, then string) *models.RecordConfig {
 	scode := strconv.Itoa(int(code))
 	rdata, err := models.ParseCFSINGLEREDIRECT([]string{name, scode, when, then}, "**current-domain**")
