+++ models/generated_types_gen.go
@@ -1,0 +1,3 @@
+package models
+
+//go:generate go run ../build/autotype
