+++ models/types.go
@@ -3,22 +3,8 @@ package models
 import (
 	"fmt"
 	"maps"
-	"strconv"
 	"strings"
-
-	"github.com/StackExchange/dnscontrol/v4/pkg/fieldtypes"
 )
-
-// RecordType is a constraint for DNS records.
-type RecordType interface {
-	A | MX | SRV | CFSINGLEREDIRECT
-}
-
-func init() {
-	MustRegisterType("A", RegisterOpts{PopulateFromRaw: PopulateFromRawA})
-	MustRegisterType("MX", RegisterOpts{PopulateFromRaw: PopulateFromRawMX})
-	MustRegisterType("SRV", RegisterOpts{PopulateFromRaw: PopulateFromRawSRV})
-}
 
 func RecordUpdateFields[T RecordType](rc *RecordConfig, rdata T, meta map[string]string) error {
 	rc.Fields = &rdata
@@ -94,194 +80,183 @@ func errorCheckFieldCount(rawfields []string, expected int) bool {
 	return len(rawfields) != expected
 }
 
-//// A
+// //// A
 
-// A is the fields needed to store a DNS record of type A
-type A struct {
-	A fieldtypes.IPv4
-}
+// func ParseA(rawfields []string, origin string) (A, error) {
 
-func ParseA(rawfields []string, origin string) (A, error) {
+// 	// Error checking
+// 	if errorCheckFieldCount(rawfields, 1) {
+// 		return A{}, fmt.Errorf("rtype A wants %d field(s), found %d: %+v", 1, len(rawfields), rawfields)
+// 	}
 
-	// Error checking
-	if errorCheckFieldCount(rawfields, 1) {
-		return A{}, fmt.Errorf("rtype A wants %d field(s), found %d: %+v", 1, len(rawfields), rawfields)
-	}
+// 	var a fieldtypes.IPv4
+// 	var err error
+// 	if a, err = fieldtypes.ParseIPv4(rawfields[0]); err != nil {
+// 		return A{}, err
+// 	}
+// 	return A{A: a}, nil
+// }
 
-	var a fieldtypes.IPv4
-	var err error
-	if a, err = fieldtypes.ParseIPv4(rawfields[0]); err != nil {
-		return A{}, err
-	}
-	return A{A: a}, nil
-}
+// // PopulateFromRawA updates rc to be an A record with contents from rawfields, meta and origin.
+// func PopulateFromRawA(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
+// 	rc.Type = "A"
 
-// PopulateFromRawA updates rc to be an A record with contents from rawfields, meta and origin.
-func PopulateFromRawA(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
-	rc.Type = "A"
+// 	// First rawfield is the label.
+// 	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
+// 		return err
+// 	}
 
-	// First rawfield is the label.
-	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
-		return err
-	}
+// 	// Parse the remaining fields.
+// 	rdata, err := ParseA(rawfields[1:], origin)
+// 	if err != nil {
+// 		return err
+// 	}
 
-	// Parse the remaining fields.
-	rdata, err := ParseA(rawfields[1:], origin)
-	if err != nil {
-		return err
-	}
+// 	return RecordUpdateFields(rc, rdata, meta)
+// }
 
-	return RecordUpdateFields(rc, rdata, meta)
-}
+// // AsA returns rc.Fields as an A struct.
+// func (rc *RecordConfig) AsA() *A {
+// 	return rc.Fields.(*A)
+// }
 
-// AsA returns rc.Fields as an A struct.
-func (rc *RecordConfig) AsA() *A {
-	return rc.Fields.(*A)
-}
+// // GetFieldsA returns rc.Fields as individual typed values.
+// func (rc *RecordConfig) GetFieldsA() fieldtypes.IPv4 {
+// 	n := rc.AsA()
+// 	return n.A
+// }
 
-// GetFieldsA returns rc.Fields as individual typed values.
-func (rc *RecordConfig) GetFieldsA() fieldtypes.IPv4 {
-	n := rc.AsA()
-	return n.A
-}
+// // GetFieldsAsStringsA returns rc.Fields as individual strings.
+// func (rc *RecordConfig) GetFieldsAsStringsA() string {
+// 	n := rc.AsA()
+// 	return n.A.String()
+// }
 
-// GetFieldsAsStringsA returns rc.Fields as individual strings.
-func (rc *RecordConfig) GetFieldsAsStringsA() string {
-	n := rc.AsA()
-	return n.A.String()
-}
+// //// MX
 
-//// MX
+// func ParseMX(rawfields []string, origin string) (MX, error) {
 
-// MX is the fields needed to store a DNS record of type MX
-type MX struct {
-	Preference uint16
-	Mx         string
-}
+// 	// Error checking
+// 	if errorCheckFieldCount(rawfields, 2) {
+// 		return MX{}, fmt.Errorf("rtype MX wants %d field(s), found %d: %+v", 1, len(rawfields)-1, rawfields[1:])
+// 	}
 
-func ParseMX(rawfields []string, origin string) (MX, error) {
+// 	var preference uint16
+// 	var mx string
+// 	var err error
+// 	if preference, err = fieldtypes.ParseUint16(rawfields[0]); err != nil {
+// 		return MX{}, err
+// 	}
+// 	if mx, err = fieldtypes.ParseHostnameDot(rawfields[1], "", origin); err != nil {
+// 		return MX{}, err
+// 	}
+// 	return MX{Preference: preference, Mx: mx}, nil
+// }
 
-	// Error checking
-	if errorCheckFieldCount(rawfields, 2) {
-		return MX{}, fmt.Errorf("rtype MX wants %d field(s), found %d: %+v", 1, len(rawfields)-1, rawfields[1:])
-	}
+// // PopulateFromRawMX updates rc to be an MX record with contents from rawfields, meta and origin.
+// func PopulateFromRawMX(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
+// 	rc.Type = "MX"
+// 	var err error
 
-	var preference uint16
-	var mx string
-	var err error
-	if preference, err = fieldtypes.ParseUint16(rawfields[0]); err != nil {
-		return MX{}, err
-	}
-	if mx, err = fieldtypes.ParseHostnameDot(rawfields[1], "", origin); err != nil {
-		return MX{}, err
-	}
-	return MX{Preference: preference, Mx: mx}, nil
-}
+// 	// First rawfield is the label.
+// 	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
+// 		return err
+// 	}
 
-// PopulateFromRawMX updates rc to be an MX record with contents from rawfields, meta and origin.
-func PopulateFromRawMX(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
-	rc.Type = "MX"
-	var err error
+// 	// Parse the remaining fields.
+// 	rdata, err := ParseMX(rawfields[1:], origin)
+// 	if err != nil {
+// 		return err
+// 	}
 
-	// First rawfield is the label.
-	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
-		return err
-	}
+// 	return RecordUpdateFields(rc, rdata, meta)
+// }
 
-	// Parse the remaining fields.
-	rdata, err := ParseMX(rawfields[1:], origin)
-	if err != nil {
-		return err
-	}
+// // AsMX returns rc.Fields as an MX struct.
+// func (rc *RecordConfig) AsMX() *MX {
+// 	return rc.Fields.(*MX)
+// }
 
-	return RecordUpdateFields(rc, rdata, meta)
-}
+// // GetFieldsMX returns rc.Fields as individual typed values.
+// func (rc *RecordConfig) GetFieldsMX() (uint16, string) {
+// 	n := rc.AsMX()
+// 	return n.Preference, n.Mx
+// }
 
-// AsMX returns rc.Fields as an MX struct.
-func (rc *RecordConfig) AsMX() *MX {
-	return rc.Fields.(*MX)
-}
+// // GetFieldsAsStringsMX returns rc.Fields as individual strings.
+// func (rc *RecordConfig) GetFieldsAsStringsMX() [2]string {
+// 	n := rc.AsMX()
+// 	return [2]string{strconv.Itoa(int(n.Preference)), n.Mx}
+// }
 
-// GetFieldsMX returns rc.Fields as individual typed values.
-func (rc *RecordConfig) GetFieldsMX() (uint16, string) {
-	n := rc.AsMX()
-	return n.Preference, n.Mx
-}
+// //// SRV
 
-// GetFieldsAsStringsMX returns rc.Fields as individual strings.
-func (rc *RecordConfig) GetFieldsAsStringsMX() [2]string {
-	n := rc.AsMX()
-	return [2]string{strconv.Itoa(int(n.Preference)), n.Mx}
-}
+// // // SRV is the fields needed to store a DNS record of type SRV
+// // type SRV struct {
+// // 	Priority uint16 `json:"priority"`
+// // 	Weight   uint16 `json:"weight"`
+// // 	Port     uint16 `json:"port"`
+// // 	Target   string `json:"target"`
+// // }
 
-//// SRV
+// func ParseSRV(rawfields []string, origin string) (SRV, error) {
 
-// SRV is the fields needed to store a DNS record of type SRV
-type SRV struct {
-	Priority uint16 `json:"priority"`
-	Weight   uint16 `json:"weight"`
-	Port     uint16 `json:"port"`
-	Target   string `json:"target"`
-}
+// 	// Error checking
+// 	if errorCheckFieldCount(rawfields, 4) {
+// 		return SRV{}, fmt.Errorf("rtype SRV wants %d field(s), found %d: %+v", 4, len(rawfields)-1, rawfields[1:])
+// 	}
 
-func ParseSRV(rawfields []string, origin string) (SRV, error) {
+// 	var priority, weight, port uint16
+// 	var target string
+// 	var err error
+// 	if priority, err = fieldtypes.ParseUint16(rawfields[0]); err != nil {
+// 		return SRV{}, err
+// 	}
+// 	if weight, err = fieldtypes.ParseUint16(rawfields[1]); err != nil {
+// 		return SRV{}, err
+// 	}
+// 	if port, err = fieldtypes.ParseUint16(rawfields[2]); err != nil {
+// 		return SRV{}, err
+// 	}
+// 	if target, err = fieldtypes.ParseHostnameDot(rawfields[3], "", origin); err != nil {
+// 		return SRV{}, err
+// 	}
+// 	return SRV{Priority: priority, Weight: weight, Port: port, Target: target}, nil
+// }
 
-	// Error checking
-	if errorCheckFieldCount(rawfields, 4) {
-		return SRV{}, fmt.Errorf("rtype SRV wants %d field(s), found %d: %+v", 4, len(rawfields)-1, rawfields[1:])
-	}
+// // PopulateFromRawSRV updates rc to be an SRV record with contents from rawfields, meta and origin.
+// func PopulateFromRawSRV(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
+// 	rc.Type = "SRV"
+// 	var err error
 
-	var priority, weight, port uint16
-	var target string
-	var err error
-	if priority, err = fieldtypes.ParseUint16(rawfields[0]); err != nil {
-		return SRV{}, err
-	}
-	if weight, err = fieldtypes.ParseUint16(rawfields[1]); err != nil {
-		return SRV{}, err
-	}
-	if port, err = fieldtypes.ParseUint16(rawfields[2]); err != nil {
-		return SRV{}, err
-	}
-	if target, err = fieldtypes.ParseHostnameDot(rawfields[3], "", origin); err != nil {
-		return SRV{}, err
-	}
-	return SRV{Priority: priority, Weight: weight, Port: port, Target: target}, nil
-}
+// 	// First rawfield is the label.
+// 	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
+// 		return err
+// 	}
 
-// PopulateFromRawSRV updates rc to be an SRV record with contents from rawfields, meta and origin.
-func PopulateFromRawSRV(rc *RecordConfig, rawfields []string, meta map[string]string, origin string) error {
-	rc.Type = "SRV"
-	var err error
+// 	// Parse the remaining fields.
+// 	rdata, err := ParseSRV(rawfields[1:], origin)
+// 	if err != nil {
+// 		return err
+// 	}
 
-	// First rawfield is the label.
-	if err := rc.SetLabel3(rawfields[0], rc.SubDomain, origin); err != nil {
-		return err
-	}
+// 	return RecordUpdateFields(rc, rdata, meta)
 
-	// Parse the remaining fields.
-	rdata, err := ParseSRV(rawfields[1:], origin)
-	if err != nil {
-		return err
-	}
+// }
 
-	return RecordUpdateFields(rc, rdata, meta)
+// // AsSRV returns rc.Fields as an SRV struct.
+// func (rc *RecordConfig) AsSRV() *SRV {
+// 	return rc.Fields.(*SRV)
+// }
 
-}
+// // GetFieldsSRV returns rc.Fields as individual typed values.
+// func (rc *RecordConfig) GetFieldsSRV() (uint16, uint16, uint16, string) {
+// 	n := rc.AsSRV()
+// 	return n.Priority, n.Weight, n.Port, n.Target
+// }
 
-// AsSRV returns rc.Fields as an SRV struct.
-func (rc *RecordConfig) AsSRV() *SRV {
-	return rc.Fields.(*SRV)
-}
-
-// GetFieldsSRV returns rc.Fields as individual typed values.
-func (rc *RecordConfig) GetFieldsSRV() (uint16, uint16, uint16, string) {
-	n := rc.AsSRV()
-	return n.Priority, n.Weight, n.Port, n.Target
-}
-
-// GetFieldsAsStringsSRV returns rc.Fields as individual strings.
-func (rc *RecordConfig) GetFieldsAsStringsSRV() [4]string {
-	n := rc.AsSRV()
-	return [4]string{strconv.Itoa(int(n.Priority)), strconv.Itoa(int(n.Weight)), strconv.Itoa(int(n.Port)), n.Target}
-}
+// // GetFieldsAsStringsSRV returns rc.Fields as individual strings.
+// func (rc *RecordConfig) GetFieldsAsStringsSRV() [4]string {
+// 	n := rc.AsSRV()
+// 	return [4]string{strconv.Itoa(int(n.Priority)), strconv.Itoa(int(n.Weight)), strconv.Itoa(int(n.Port)), n.Target}
+// }
